+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 Common:
 * Added dataset loader for Babel dataset.
 
+Task 1, Predict Modality:
+* Starter Kit: Added starter kit for Python users
+
+Task 3, Joint Embedding:
+* Metric: Added TotalVI metric 'latent_mixing_metric'.
+
 # neurips2021_multimodal_viash 0.4.0
 
 ## NEW FEATURES